+++ src/Graphics/Image/ColorSpace/HSI.hs
@@ -186,7 +186,7 @@ instance Foldable (Pixel HSIA) where
 
 
 instance Storable e => Storable (Pixel HSIA e) where
-  sizeOf _ = 3 * sizeOf (undefined :: e)
+  sizeOf _ = 4 * sizeOf (undefined :: e)
   alignment _ = alignment (undefined :: e)
   peek p = do
     q <- return $ castPtr p