+++ src/Graphics/Image/ColorSpace/YCbCr.hs
@@ -189,7 +189,7 @@ instance Foldable (Pixel YCbCrA) where
 
 instance Storable e => Storable (Pixel YCbCrA e) where
 
-  sizeOf _ = 3 * sizeOf (undefined :: e)
+  sizeOf _ = 4 * sizeOf (undefined :: e)
   alignment _ = alignment (undefined :: e)
   peek p = do
     q <- return $ castPtr p